# resources/swagger/swagger.yml
info:
  title: Comentario
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
swagger: "2.0"
basePath: /api
securityDefinitions:
  # Header authentication for commenters (used by the embed part)
  userSessionHeader:
    type: apiKey
    in: header
    name: X-User-Session

  # Cookie authentication for users (used by the frontend part). Uses the apiKey type for the lack of a proper
  # cookie-based authentication in Swagger 2
  userCookie:
    type: apiKey
    in: header
    name: Cookie

  # Token authentication for certain endpoints, called externally, such as password reset or email confirmation
  token:
    type: oauth2
    flow: accessCode
    authorizationUrl: http://dummy/
    tokenUrl: http://dummy/
    scopes:
      confirm-email: confirm user's email
      login: authenticate the user
      pwd-reset: reset user's password

# Default security is cookie-based user authentication
security:
  - userCookie: []
definitions:
...
  commentSort:
    description: Comment sorting. 1st letter defines the property, 2nd letter the direction
    type: string
    minLength: 2
    maxLength: 2
    enum:
      - ta # By timestamp, ascending
      - td # By timestamp, descending
      - sa # By score, ascending
      - sd # By score, descending
    x-isnullable: false
  domain:
    description: Registered domain
    type: object
    required:
      - host
      - isHttps
    properties:
      id:
        type: string
        format: uuid
        readOnly: true
        description: Unique domain ID
      name:
        type: string
        maxLength: 255
        description: Domain display name
      host:
        $ref: "#/definitions/host"
        description: Domain host
        x-isnullable: false
      createdTime:
        type: string
        format: date-time
        readOnly: true
        description: When the domain was created
      isHttps:
        type: boolean
        description: Whether HTTPS should be used to resolve URLs on this domain (as opposed to HTTP)
        x-omitempty: false
      isReadonly:
        type: boolean
        readOnly: true
        description: Whether the domain is readonly (no new comments are allowed)
        x-omitempty: false
        x-isnullable: false
      authAnonymous:
        type: boolean
        description: Whether commenting by unregistered users is allowed
        x-omitempty: false
      authLocal:
        type: boolean
        description: Whether local authentication is allowed
        x-omitempty: false
      authSso:
        type: boolean
        description: Whether SSO authentication is allowed
        x-omitempty: false
      ssoUrl:
        type: string
        description: SSO provider URL
        x-omitempty: false
      ssoNonInteractive:
        type: boolean
        description: Whether to use a non-interactive SSO login
        x-omitempty: false
      ssoSecretConfigured:
        type: boolean
        readOnly: true
        description: Whether SSO secret is configured for this domain
        x-omitempty: false
        x-isnullable: false
      modAnonymous:
        type: boolean
        description: Whether all comments by unregistered users are to be approved by a moderator
        x-omitempty: false
      modAuthenticated:
        type: boolean
        description: Whether all comments by authenticated users are to be approved by a moderator
        x-omitempty: false
      modNumComments:
        type: integer
        format: uint
        description: Number of first comments by user on this domain that require a moderator approval
        x-omitempty: false
      modUserAgeDays:
        type: integer
        format: uint
        description: Number of first days since user has registered on this domain to require a moderator approval on their comments
        x-omitempty: false
      modLinks:
        type: boolean
        description: Whether all comments containing a link are to be approved by a moderator
        x-omitempty: false
      modImages:
        type: boolean
        description: Whether all comments containing an image are to be approved by a moderator
        x-omitempty: false
      modNotifyPolicy:
        $ref: "#/definitions/domainModNotifyPolicy"
        description: Moderator notification policy for domain
      defaultSort:
        $ref: "#/definitions/commentSort"
        description: Default comment sorting for domain
      countComments:
        type: integer
        readOnly: true
        description: Total number of comments. -1 means the value is not provided
        x-omitempty: false
      countViews:
        type: integer
        readOnly: true
        description: Total number of views. -1 means the value is not provided
        x-omitempty: false
      rootUrl:
        type: string
        format: uri
        description: Root URL of the domain, without trailing slash
        readOnly: true
  domainExtension:
    description: Domain extension info
    type: object
    required:
      - id
    properties:
      id:
        $ref: "#/definitions/domainExtensionId"
        description: Extension ID
        example: akismet
      name:
        type: string
        readOnly: true
        description: Extension display name
        example: Akismet
      config:
        type: string
        description: Extension configuration, linebreak-separated list of key=value pairs
      requiresKey:
        type: boolean
        readOnly: true
        description: Whether the extension requires an API key at domain level
        x-omitempty: false
  domainExtensionId:
    description: Domain extension ID
    type: string
    enum:
      - akismet
      - perspective
      - apiLayer.spamChecker
    x-isnullable: false
  domainModNotifyPolicy:
    description: Moderator notification policy for domain
    type: string
    enum:
      - none
      - pending
      - all
    x-isnullable: false
...
  domainUser:
    description: Registered user on a domain
    type: object
    readOnly: true
    properties:
      domainId:
        type: string
        format: uuid
        description: ID of the domain
      userId:
        type: string
        format: uuid
        description: ID of the user
      isOwner:
        type: boolean
        description: Whether the user is an owner of the domain (assumes is_moderator and is_commenter)
        x-omitempty: false
      isModerator:
        type: boolean
        description: Whether the user is a moderator of the domain (assumes is_commenter)
        x-omitempty: false
      isCommenter:
        type: boolean
        description: Whether the user is a commenter of the domain (if false, the user is readonly on the domain)
        x-omitempty: false
      notifyReplies:
        type: boolean
        description: Whether the user is to be notified about replies to their comments
        x-omitempty: false
      notifyModerator:
        type: boolean
        description: Whether the user is to receive moderator notifications (only when is_moderator is true)
        x-omitempty: false
      notifyCommentStatus:
        type: boolean
        description: Whether the user is to be notified about status changes (approved/rejected) of their comments
        x-omitempty: false
      createdTime:
        type: string
        format: date-time
        description: When the domain user was created
  dynamicConfigItem:
    description: Dynamic configuration item
    type: object
    required:
      - key
      - value
    properties:
      key:
        type: string
        maxLength: 255
        description: Item key
      value:
        type: string
        maxLength: 255
        description: Item value
      datatype:
        $ref: "#/definitions/dynamicConfigItemDatatype"
        readOnly: true
        description: Item datatype
      updatedTime:
        type: string
        format: date-time
        readOnly: true
        description: Timestamp when the item was last updated in the database
      userUpdated:
        type: string
        format: uuid
        readOnly: true
        description: Reference to the user who last updated the item in the database
      defaultValue:
        type: string
        readOnly: true
        description: Item's default value
      section:
        type: string
        description: Key of the section the item belongs to
      min:
        type: integer
        readOnly: true
        description: Minimum allowed value of the setting
        x-omitempty: false
      max:
        type: integer
        readOnly: true
        description: Maximum allowed value of the setting
        x-omitempty: false
  dynamicConfigItemDatatype:
    description: Dynamic configuration item datatype
    type: string
    enum:
      - bool
      - int
...
  federatedIdpId:
    description: Federated identity provider ID
    type: string
    maxLength: 37
    pattern: '^facebook|github|gitlab|google|twitter|(oidc:[-a-z0-9]{1,32})$'
    x-isnullable: false
  host:
    description: DNS hostname, with an optional port number
    example: example.com:8080
    type: string
    minLength: 1
    maxLength: 259  # 253 (max length of a DNS name) + ":" + 5 (port number 1..65535)
    pattern: "[-.a-z0-9]{1,253}(:[0-9]{1-5})?"
    x-isnullable: false
...
  importResult:
    description: Comment import result
    type: object
    readOnly: true
    properties:
      usersTotal:
        type: integer
        format: uint
        description: Total number of users
        x-omitempty: false
      usersAdded:
        type: integer
        format: uint
        description: Number of added users
        x-omitempty: false
      domainUsersAdded:
        type: integer
        format: uint
        description: Number of added domain users
        x-omitempty: false
      pagesTotal:
        type: integer
        format: uint
        description: Total number of domain pages
        x-omitempty: false
      pagesAdded:
        type: integer
        format: uint
        description: Number of added domain pages
        x-omitempty: false
      commentsTotal:
        type: integer
        format: uint
        description: Total number of comments processed
        x-omitempty: false
      commentsImported:
        type: integer
        format: uint
        description: Number of imported comments
        x-omitempty: false
      commentsSkipped:
        type: integer
        format: uint
        description: Number of skipped comments
        x-omitempty: false
      commentsNonDeleted:
        type: integer
        format: uint
        description: Number of non-deleted imported comments
        x-omitempty: false
      error:
        type: string
        description: Any error message occurred during the import
...
parameters:
...
  pathImportSource:
    name: source
    in: path
    required: true
    description: Source of the data to import into a domain
    type: string
    enum:
      - comentario
      - disqus
      - wordpress
  pathUuid:
    in: path
    name: uuid
    required: true
    description: UUID in the path
    type: string
    format: uuid
    x-isnullable: false
  queryDomainId:
    in: query
    name: domain
    required: true
    description: Domain UUID in the query
    type: string
    format: uuid
  queryFilter:
    name: filter
    in: query
    required: false
    description: Optional substring to filter the result by
    type: string
    maxLength: 100
...
  querySortDesc:
    name: sortDesc
    in: query
    required: false
    description: 'Optional sort direction: ascending (false) or descending (true)'
    type: boolean
  queryPageNumber:
    name: page
    in: query
    required: false
    description: Optional page number to return results for
    type: integer
    format: uint
    minimum: 1
    default: 1
...
paths:
...
  /domains:
    get:
      operationId: DomainList
      summary: Get a list of registered domains
      tags:
        - ApiGeneral
      parameters:
        - $ref: "#/parameters/queryFilter"
        - $ref: "#/parameters/queryPageNumber"
        - in: query
          name: sortBy
          type: string
          enum:
            - name
            - host
            - created
            - countComments
            - countViews
          description: Property to sort results by
        - $ref: "#/parameters/querySortDesc"
      responses:
        200:
          description: List of domains
          schema:
            type: object
            properties:
              domains:
                type: array
                items:
                  $ref: "#/definitions/domain"
                description: Domains the current user has (some) rights to
              domainUsers:
                type: array
                items:
                  $ref: "#/definitions/domainUser"
                description: Current user's properties on domains (can be linked to a domain using domainId)

    post:
      operationId: DomainNew
      summary: Register a new domain
      tags:
        - ApiGeneral
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - domain
            properties:
              domain:
                $ref: "#/definitions/domain"
                description: Domain properties
              configuration:
                type: array
                items:
                  $ref: "#/definitions/dynamicConfigItem"
                description: Domain configuration items
              federatedIdpIds:
                type: array
                items:
                  $ref: "#/definitions/federatedIdpId"
                description: IDs of federated identity providers enabled for the domain
              extensions:
                type: array
                items:
                  $ref: "#/definitions/domainExtension"
                description: List of extensions enabled for the domain, and their configurations
      responses:
        200:
          description: Domain added successfully
          schema:
            $ref: "#/definitions/domain"
            description: The added domain
  /domains/count:
    get:
      operationId: DomainCount
      summary: Get number of domains available to the current user
      tags:
        - ApiGeneral
      parameters:
        - in: query
          name: owner
          type: boolean
          description: >
            Only include domains where the current user is an owner. Ignored if moderator is true because every owner is
            also a moderator
        - in: query
          name: moderator
          type: boolean
          description: Only include domains where the current user is a moderator
      responses:
        200:
          description: Number of domains
          schema:
            type: integer
  /domains/{uuid}:
    parameters:
      - $ref: "#/parameters/pathUuid"

    get:
      operationId: DomainGet
      summary: Get properties of a domain
      tags:
        - ApiGeneral
      responses:
        200:
          description: Domain properties
          schema:
            type: object
            properties:
              domain:
                $ref: "#/definitions/domain"
                description: Domain properties
              domainUser:
                $ref: "#/definitions/domainUser"
                description: Current user's properties on the domain (only when there is one)
              configuration:
                type: array
                items:
                  $ref: "#/definitions/dynamicConfigItem"
                description: Domain configuration items
              federatedIdpIds:
                type: array
                items:
                  $ref: "#/definitions/federatedIdpId"
                description: IDs of federated identity providers enabled for the domain
              extensions:
                type: array
                items:
                  $ref: "#/definitions/domainExtension"
                description: List of extensions enabled for the domain, and their configurations

    put:
      operationId: DomainUpdate
      summary: Update properties of specified domain
      tags:
        - ApiGeneral
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - domain
            properties:
              domain:
                $ref: "#/definitions/domain"
                description: Domain properties
              configuration:
                type: array
                items:
                  $ref: "#/definitions/dynamicConfigItem"
                description: Domain configuration items
              federatedIdpIds:
                type: array
                items:
                  $ref: "#/definitions/federatedIdpId"
                description: IDs of federated identity providers enabled for the domain
              extensions:
                type: array
                items:
                  $ref: "#/definitions/domainExtension"
                description: List of extensions enabled for the domain, and their configurations
      responses:
        200:
          description: Domain properties have been updated
          schema:
            $ref: "#/definitions/domain"
            description: The updated domain

    delete:
      operationId: DomainDelete
      summary: Delete a domain
      tags:
        - ApiGeneral
      responses:
        204:
          description: Domain has been deleted
  /domains/{uuid}/clear:
    delete:
      operationId: DomainClear
      summary: Clear all domain's pages/comments/votes/views
      tags:
        - ApiGeneral
      parameters:
        - $ref: "#/parameters/pathUuid"
      responses:
        204:
          description: Domain has been cleared
  /domains/{uuid}/purge:
    delete:
      operationId: DomainPurge
      summary: Permanently remove all deleted domain's comments and/or comments by deleted users
      tags:
        - ApiGeneral
      parameters:
        - $ref: "#/parameters/pathUuid"
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              markedDeleted:
                type: boolean
                description: Whether to remove comments marked as deleted
              userCreatedDeleted:
                type: boolean
                description: Whether to remove comments written by users who are now deleted
      responses:
        200:
          description: Domain has been purged
          schema:
            type: object
            required:
              - commentCount
            properties:
              commentCount:
                type: integer
                description: Total number of deleted comments
                x-isnullable: false
  /domains/{uuid}/readonly:
    put:
      operationId: DomainReadonly
      summary: Set the domain's readonly state
      tags:
        - ApiGeneral
      parameters:
        - $ref: "#/parameters/pathUuid"
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - readonly
            properties:
              readonly:
                type: boolean
                description: Readonly status to set
      responses:
        204:
          description: Domain status has been set
  /domains/{uuid}/export:
    get:
      operationId: DomainExport
      summary: Export domain data and download as a gzip-archive file
      tags:
        - ApiGeneral
      produces:
        - application/gzip
      parameters:
        - $ref: "#/parameters/pathUuid"
      responses:
        200:
          description: Export file
          schema:
            type: file
          headers:
            Content-Disposition:
              type: string
  /domains/{uuid}/import/{source}:
    post:
      operationId: DomainImport
      summary: Import domain data (commenters, pages, comments) from a data dump
      tags:
        - ApiGeneral
      consumes:
        - multipart/form-data
      parameters:
        - $ref: "#/parameters/pathUuid"
        - $ref: "#/parameters/pathImportSource"
        - in: formData
          name: data
          type: file
          maxLength: 10485760 # 10 MiB
          required: true
          description: Import data file
      responses:
        200:
          description: Import completed successfully
          schema:
            $ref: "#/definitions/importResult"
  /domains/{uuid}/sso/new:
    post:
      operationId: DomainSsoSecretNew
      summary: Generate an SSO secret for specified domain
      tags:
        - ApiGeneral
      parameters:
        - $ref: "#/parameters/pathUuid"
      responses:
        200:
          description: SSO secret has been generated for the domain
          schema:
            type: object
            properties:
              ssoSecret:
                type: string

  #---------------------------------------------------------------------------------------------------------------------
  # Domain pages
  #---------------------------------------------------------------------------------------------------------------------
...
  /comments/count:
    get:
      operationId: CommentCount
      summary: Get the number of comments
      tags:
        - ApiGeneral
      parameters:
        - $ref: "#/parameters/queryDomainId"
        - in: query
          name: pageId
          required: false
          description: Optional domain page ID to filter comments by
          type: string
          format: uuid
        - in: query
          name: userId
          required: false
          description: Optional user ID to filter comments by
          type: string
          format: uuid
        - in: query
          name: approved
          type: boolean
          required: false
          description: Whether to include approved comments
        - in: query
          name: pending
          type: boolean
          required: false
          description: Whether to include comments pending moderation
        - in: query
          name: rejected
          type: boolean
          required: false
          description: Whether to include rejected comments
        - in: query
          name: deleted
          type: boolean
          required: false
          description: Whether to include deleted comments
      responses:
        200:
          description: Number of comments matching the criteria
          schema:
            type: integer
            x-omitempty: false
...